chore: rename warmup to download_model
